feat(inventory): migrate stock quantities from integer to decimal
Support fractional stock end-to-end (e.g. a recipe consuming 0.5 L no
longer truncates to 0/1). Changed to float64/double precision:
- InventoryBalance.on_hand/available/reserved
- Reservation items JSON (requested/reserved/available qty)
- purchase_order_lines, stock_transfer_lines, inventory_lots,
  goods_receipt_lines, requisition_lines, rfq_lines, rfq_awards quantities

Also removed the int truncations in BOM explosion, consumption deduct,
restock, and bulk-import opening balance. reorder_level/reorder_quantity
stay integer. DTOs, stock-service math, low-stock checks, analytics, and
the auto-reorder consumer updated accordingly. Atlas migration
20260605075831 (bigint -> double precision, in-place).

Cross-service note: ordering-backend already parses these as float64
(commit d1fdac5); pos-api was already decimal-ready.

+++ cmd/seed/seed_balances.go
@@ -96,8 +96,8 @@ func seedBalances(ctx context.Context, client *ent.Client, tenantID uuid.UUID, s
 			SetTenantID(tenantID).
 			SetItemID(itm.ID).
 			SetWarehouseID(wh.ID).
-			SetOnHand(onHand).
-			SetAvailable(onHand).
+			SetOnHand(float64(onHand)).
+			SetAvailable(float64(onHand)).
 			SetReserved(0).
 			SetReorderLevel(reorderLvl).
 			SetReorderQuantity(reorderLvl).
